[Fix] city arg sent may not match with the one Yahoo returns
City argument for Yahoo weather api may not match with the location data it sends back. This may not be a issue when you determine location using IP address , but when user inputs city himself using --city and it doesn't match with Yahoo database, it may be difficult to pinpoint the error.
This commit enables it to print location returned by Yahoo Weather API, instead of city mentioned earlier.